requirements.txt
python-dotenv
pyunsplash
twikit
schedule
Pillow